% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_shp_files}
\alias{compare_shp_files}
\title{compare_shp_files}
\usage{
compare_shp_files(dbx_shp, new_shp)
}
\arguments{
\item{dbx_shp}{the last Dropbox version}

\item{new_shp}{the update version}
}
\value{
a list with $summary and $full_comp. summary contains the summary of
the comparison as character and full_comp the full comparison as tibble
}
\description{
This function compares two versions of the same crowns vector data. Vectors as
to be sf object and the name for the 'id', 'idtax_f' and 'geometry' have to be
'id', 'idtax_f' and 'geometry'. The function then return a list with the summary
of the comparison as character and the full comparison as tibble.
}
\examples{
st_geometry(dbx_shp) <- "geometry"
st_geometry(new_shp) <- "geometry"
compare_shp_files(dbx_shp, new_shp)

}
