% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_module.R
\name{comparison_server}
\alias{comparison_server}
\title{Server module for comparing two shapefiles}
\usage{
comparison_server(input, output, session, dbx_shp, new_shp)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{dbx_shp}{sf object. Reference shapefile.}

\item{new_shp}{sf object. New shapefile to compare.}
}
\value{
A list containing the comparison results.
}
\description{
Compares two sf objects (shapefiles) and provides outputs for a summary
and a detailed comparison table using DT.
}
