% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_module.R
\name{comparison_ui}
\alias{comparison_ui}
\title{UI module for comparing two shapefiles}
\usage{
comparison_ui(id)
}
\arguments{
\item{id}{Character. Shiny module id.}
}
\value{
A Shiny UI element (`tagList`) to be used in a Shiny app.
}
\description{
Provides a Shiny UI for comparing two shapefiles, showing a summary and
a detailed comparison table.
}
