% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correct_id.R
\name{get_correct_id}
\alias{get_correct_id}
\title{Correct the new file from the id find in the old version}
\usage{
get_correct_id(new_shp, old_shp, crs)
}
\arguments{
\item{new_shp}{blabla}

\item{old_shp}{blabla}

\item{crs}{blabla}
}
\description{
blabla
}
