% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{plot_ui}
\alias{plot_ui}
\title{UI module for interactive map of polygons}
\usage{
plot_ui(id)
}
\arguments{
\item{id}{Character. Shiny module id.}
}
\value{
A Shiny UI object (tabPanel with leafletOutput).
}
\description{
This function creates the UI part of a Shiny module displaying polygons on a Leaflet map.
}
