% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_module.R
\name{update_server}
\alias{update_server}
\title{Server logic for update module}
\usage{
update_server(input, output, session, site, new_shp, comparison)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{site}{Character. Name of the site to use in output file naming}

\item{new_shp}{`sf` object. New shapefile with updated crowns}

\item{comparison}{`sf` object. Comparison data for crowns}
}
\value{
None. Provides reactive download via Shiny.
}
\description{
Server logic for `update_ui`. Handles reading uploaded files, updating comparison data,
and providing a downloadable updated file.
}
