% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_module.R
\name{update_ui}
\alias{update_ui}
\title{Update UI module for comparison of crown data}
\usage{
update_ui(id)
}
\arguments{
\item{id}{Character. Module namespace id.}
}
\value{
A Shiny UI tagList.
}
\description{
This Shiny module creates a UI to upload old comparison files, update data on Dropbox,
and download the updated comparison dataset.
}
