% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_dbx_data_module.R
\name{upload_dbx_data_server}
\alias{upload_dbx_data_server}
\title{Server module for uploading a Dbx shapefile}
\usage{
upload_dbx_data_server(input, output, session)
}
\arguments{
\item{input}{Shiny input}

\item{output}{Shiny output}

\item{session}{Shiny session}
}
\value{
A reactive list containing the shapefile (`shp`) and selected site (`site`) when checkbox is checked
}
\description{
Server module for uploading a Dbx shapefile
}
