% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_dbx_data_module.R
\name{upload_dbx_data_ui}
\alias{upload_dbx_data_ui}
\title{UI module for uploading a Dbx shapefile}
\usage{
upload_dbx_data_ui(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
Shiny UI elements for uploading a shapefile and selecting a column
}
\description{
UI module for uploading a Dbx shapefile
}
