% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_updated_data_module.R
\name{upload_updated_data_server}
\alias{upload_updated_data_server}
\title{Server module for uploading an updated shapefile}
\usage{
upload_updated_data_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A reactive expression returning the updated `sf` object if the checkbox is checked, or NULL otherwise.
}
\description{
This module handles the server-side logic for uploading a shapefile,
renaming a column, and returning the updated `sf` object.
}
