% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_updated_data_module.R
\name{upload_updated_data_ui}
\alias{upload_updated_data_ui}
\title{UI module for uploading an updated shapefile}
\usage{
upload_updated_data_ui(id)
}
\arguments{
\item{id}{A character string; the module's namespace ID.}
}
\value{
A Shiny UI element (tagList) for file input, selectInput, and checkbox.
}
\description{
This module provides a UI to upload a shapefile, select a column,
and confirm that the updated file is ready.
}
